Fixes #2621, migrating common code from ResourceUtil to Translator
