+++ src/main/java/genetics/api/alleles/AlleleCategorized.java
@@ -3,7 +3,7 @@ package genetics.api.alleles;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
 
-import forestry.core.utils.ResourceUtil;
+import forestry.core.utils.Translator;
 
 public class AlleleCategorized extends Allele {
 
@@ -30,6 +30,6 @@ public class AlleleCategorized extends Allele {
 	@Override
 	public ITextComponent getDisplayName() {
 		String customName = modId + '.' + "allele." + category + '.' + valueName;
-		return ResourceUtil.tryTranslate(customName, getLocalisationKey());
+		return Translator.tryTranslate(customName, getLocalisationKey());
 	}
 }
